There was a bug leading to "Segmentation fault" failure, triggered (since the commit of about May 21) by
pre_bottom_flag:=true
FPPpre(E6_q)

Turned out to be in "deform_support_int_hash" defined in unity.at;
the bug was probably very old, but the function was not invoked until
some recent commit. This was an example of a bug that led to a
function invoking a non-existent element of a list; the kind of
problem that atlas seems to catch when it's invoked "directly," but
escapes notice when buried in a few levels of function. I don't know
how to tell when atlas might have/should have caught it.

Anyway it seems to be fixed now.