Fix long standing bug in |twisted_deform|: check flip when using stored formula
Deformation formulas are stored just once for a complete "alcove", a range of
parameters that can be continuously deformed into one another without crossing
any reducibility hyperplanes (taken to include cases of entering or leaving such
a hyperplane to side where the formula changes). However, the twisted
deformation formula is also affected by flips of the default extension that
follow another regime, hard to pin down; in particular a flip may occur between
parameters in the same alcove. In case a twisted formula was needed for a
parameter and a formula was already stored for its alcove, the code used to use
this form without adapting the |flip| output according to the positions of the
current parameter and to (older) one for which the formula was stored. This is
now corrected, using essentially the recently added |shift_flip| test.