Fixed BUG in bottom_layer code
First noticed in E6_s with Levis precomputed, at x = 204, lambda =
[4,2,1,3,1,-4]; in this case, with more_shift = 0,
FPP_unitary_hash([(x,lambda)]) reports as NONUNITARY several unitary
representations. The culprit is the edge with center at nu =
[-4,3,1,1,1,-4]/6. This rep is good range cohomologically induced from
Spin(5,3); but the inducing rep is NOT the hermitian center of its
facet. The code saw that it was supposed to know all about FPP unitary
reps for Spin(5,3), looked this one up, and didn't find it (because
what was recorded was the hermitian center of the facet). Therefore
the rep was incorrectly declared nonunitary.

Fixed the is_unitary lookup functions in FPP_facets_herm.at (to
replace argument p by herm_center(p)). This seems to quash the bug.