changed ICPU and ICPNU defaults to false, fix bug apparent in hash2 (E6_s spherical)
The two ICP flags turn out to slow everything for E6_s by a factor of
maybe 5, almost all because of the case of minimal principal series.
I think this was a case of traversing an entire Weyl group to find
some (much smaller number) of coset reps; so the code ought to be
possible to speed up to usefulness. But it isn't there now, and I'm
not going to work more on it now.

There are bitmaps to record which (x,lambda) the entire FPP dual is
known. In the FPP_unitary_hash_general function, these flags were set
just BEFORE the actual computation took place. I have been modifying
the is_unitary functions to take more advantage of what is known, and
this finally bit me in the leg: is_unitary asked about a parameter
(x,lambda,nu) for the (x,lambda) currently being worked on; the
(prematurely) set flag asserted that I already knew all the unitary nu
for this (x,lambda); the code therefore looked up p(x,lambda,nu) in
the unitary hash table, failed to find it, and (incorrectly) concluded
that p was nonunitary.  I think it's fixed now.

Various minor changes that may speed things, especially for E8, but do
not affect interface. For example, in order to tell whether (x,lambda)
had been computed, the code previously had to compute all the lambdas
for x, and locate lambda in this list. For E8 this took perhaps
minutes. Often when _one_ (x,lambda) is computed, then _all_ the
(x,lambda') are computed at the same time. There is now (at Steve
Miller's suggestion) a bitmap of the x for which ALL (x,lambda) are
known. This is consulted before trying to locate lambda in its list,
which often avoids computing the list.