Provide |Free_Abelian_light<C>| with an equality test to facilitate |assert|
The method chosen does not require any new memory (as would computing the
difference between two values and testing |is_zero|) and can stop rapidly in
case a coefficient difference is found. On the downside it test the coefficients
of a same term twice in case the arguments are actually equal; for a method that
is mainly there for use in |assert| statements, this is not such a big deal.