Add experimental version of |full_deform| with time limiting argument
Here a second, integer, argument to the function gives a time limit in
milliseconds. If the call returns within the allotted time, this is equivalent
to the call without time limit argument. For now the solution when the timer
runs out instead, is to terminate whatever command was in progress, as if halted
by a user interrupt. The ultimate goal is to instead catch the condition at the
point of invocation of the function, and to have it deliver a union value that
indicates either failure (in the mentioned case) or success with a value.