Add non-tagged version of discrimination clauses (CASE .. | .. | .. ESAC).
In this version, the branches must come in the order of the variants of the
union type, which makes it superfluous to label those branches with tags; also
the identifier pattern implicitly receives the variant type of the union, so it
is not necessary to specify its type. Here in both cases "unnecessary" means
"not allowed" as well. In the current commit, there is still the possibility to
alternatively specify function expressions for each branch, using "CASE ... IN"
syntax rather than "CASE .. |"; this likely serves no more purpose and can go.
It should be noted that an advantage of the untagged discrimination expression
over the latter "union case" clause is that it allows branch expressions to
contain |return| or |break| statements into the context of the CASE clause,
which the function expressions in "union-case" would capture, and thus forbid.