Change most |type_expr| constructors into named factory methods.
Since factory functions are named, this change should improve significantly
readability, telling which kind of type is being constructed. In some places
where the name had to be inserted, it was not even clear by inspection at which
point in the expression it had to be inserted (a |type_expr| was constructed).

In passing, a parser |id_code| was changed from |short| to |unsigend short|.
and two |mk| functions were added.