Simplify the recursive descent parser of type strings: no "void" necessary
While "void" is used as a necessary type designation in user programs, the type
strings for installing built-in can simply use the notation for a 0-tuple, as no
confusion with an expression is possible. The primitive type scanning part of
the type string parser is thereby simplified. Also, the |type_ptr| returning
version |mk_type| was found unnecessary, and eliminated, and some local
|type_ptr| variables replaced by just the |type_expr| pointed to.

Finally we clarified some of the commentary in global.w.