Modify |resolve_overload|: all exact matches are tested before any inexact ones.
I don't recall why it was originally done in a single loop, testing each variant
both for exact and inexact matches before passing to the next; it probably dates
from the times that a separate conversion was attempted for each variant without
computing an a priori argument type at all. In any case the two-pass approach
now seems much cleaner, and it will facilitate handling of generic functions
without the current special-purpose code, which we intend to implement soon.

We also cleaned up the code documentation for the |resolve_overload| function.