Try to address serious slowness for small x
The problems are more or less connected to the default
interrupt_flag=true; assigning this to false is perhaps a good idea
for now? The difficulty is that I have tried to do a time-limited
version of is_unitary_to_ht, and this is more or less impossible
because (for small x?) huge amounts of time are spent in the
uninterruptible block_deform.