Enrich |type| with more methods, use them mostly in |balance| for now.
We try to develop a strategy for replacing |type_expr| by |type| in those places
where a fully determined, though possibly polymorphic, type is needed. We intend
that types coming from context, as in the type arguments to |balance| and
|convert_expr|, will remain |type_expr| values that may have undetermined parts
(often even being completely undetermined). A type will often start out as such
a |type_expr|, then be refined by calls to the |specialise| method, and when it
finally gets to become the type of an expression, being converted by
|type::wrap|. Usually there are no undetermined parts left at the final step,
but if there are (as is the case for an empty list) they get converted into
fresh type variables. Note however that when a type variable should be present
in a type pattern coming from context, it does not behave as an undetermined
type sub-expression, but as a goal that must be matched by specializing
(substituting into) the a priori type of the expression. The |broader_eq|
function used in balancing tries to cater for these distinctions.

Remove unused methods |record_match|; we use |matches| now instead.