Fix a problem with arguments declared of type [*]: disallow and remove them
The type resulting from writing [*] produces an internal type that now breaks
the type rules. It will in the future be possible, but not useful, to declare a
parameter type that only matches the empty list. The reason it will not be
useful is that the empty list will also match any other row type, and overload
resolution for argument values with polymorphic types will forbid the case of
multiple exact matches (a rule not yet implemented). So explicitly declaring a
match for just the empty list will actually break matching an argument [],
because in all likelyhood it will produce a double match. As long as only one
row type matches, an argument [] should be allowed and work in future versions.