Remove the |specialise| and other methods meant to alter existing variable types
We cannot allow (and should never have allowed) any use of a variable change the
type of the variable, even if it is an assignment that makes it a polymorphic
type into a monomorphic instance of it. It can be abused, so cannot be tolerated.

Given the effort and work-arounds that had been undertaken to implement this
"feature", this change means a major simplification of some part of the code.

This change was followed by cleaning up, which involves removing the cheating
with constness in |type_expr| methods (selecting a non-|const| field from a
|const| value); the result is that more proper use of |type_expr| is enforced.