Fix treatment of type variable as pattern in |specialise|
Having a free type variable as pattern argument of |specialise| (we should test
that it is free, but cannot yet so we simply assume it is) in general means that
the expression that gave this pattern is like a |return| or call of |error|, so
cannot ever return a value. This should allow the specialisation to always work,
but it should not preempt setting |*this| to the target type if it was
undefined. Failing to do this made balance run into problems, making the loading
of basic.at impossible.

Also add missing |make_type_variable| function