Fix bug in |type::unify|, and also forbid a weird and problematic case in it
The bug is to mistakenly use the |type_nr| method instead of |typevar_count| in
a type variable case; the forbidden case is where |type::unify| matches a type
variable to an unstable yet partially determined type pattern. The comments in
the code explain why we choose to not handle that case, of which it is unclear
whether it can arise at all.