Use unifying in going from a stored-type expression to an imposed context type
This allows for instance the value of a polymorphic identifier to be used where
a monomorphic (or less polymorphic) instance is required. Other stored-type
expressions are function calls, whether overloaded or of some other form,
array subscription and slices.

This also allowed an awkward case of |type_expr::substitute| (that of type
variable as pattern) to be removed; the case had to be treated separately to
accept for instance calls to |error| in positions that require a specific type.
Now we no longer use |specialise| in such cases, but rather a new instance of
|conform_types| using one-way unification in case of a stable required type.