Introduce |fc| argument to analyze types, and reorganise grammar a bit
Currently |fc| is |0| in every call to |analyse)types|, but this should change
in the future. The grammar reorganisation is such that operator definitions are
henceforth subsumed in 'declaration'. Also change identifier code and priority
of operator symbols to |unsigned short| (rather than just |short|), the former
to match other identifier codes.