Correct |type::matches|, and define separate method |type::matches_argument|
The method |type::matches|, when used for overloaded function calls, did not
function well if called from a context that has fixed type variables, as they
get confounded with the polymorphic variables of the function type. It is now
redesigned for the overloaded case, in which all type variables in the function
type are polymorphic, by shifting them to above any type variables that could
occur in the actual argument type. In general function calls, where the function
can be given by any kind of expression, the situation is different (the function
type can have variables fixed in the context), and deserves the definition of a
separate method to handle it, which is what |type::matches_argument| does.

This commit also makes changes to operator casts, so that they correctly look up
polymorphic entries in the overload table, and record when a match of such an
entry was found that requires substitution for some type variables in the
polymorphic type. For operator casts in a context with abstract type variables,
the lookup of an exact match was modified by introducing a shift if necessary to
avoid capturing any fixed type variables that happen to be equal to a stored
polymorphic type variable: |overload_table::entry| gets an optional parameter.