Added red_count.at...
...which counts positive roots that may contribute to reducibility of
a Param. New global flag red_count_flag in FPP_faces_herm.at,
defaulting to false; if set to true, the calculations of K-characters
check to see whether there is a boundary face having the same
red_count, and therefore same K-character.

For Spin(4,4), red_count_flag reduces the number of K-character
calculations from 2700 to 1800, and reduces memory use a bit; but
because red_count is slow, the code runs slightly slower.