Fix a subtle hole in the axis language: parameter lists could give 1-tuple types
A typical exanple is |((int n),int m):n+m| which gave type ((int),int->int),
where the (int) is not equivalent to int but is a 1-tuple with one |int| field.
The solution adopted is not to have the grammar rule this out (which turned out
to run into shift/reduce conflicts) but to remove the parentheses while parsing.