Leave type checking of |let| expressions, testing tuple structure early
This commit initially did change type checking of |let| expressions to test
tuple structure after checking the right hand side in an undetermined type
context, thinking that would facilitate handling op tuple types by not providing
a partially determined tuple type context. That turned out to be a bad idea, not
just because such type context could not be avoided anyway, but also because a
strong void type context after an identifier pattern '()' is desirable.

Therefore this commit is reduced to some cosmetic changes and insertion of an
explanantion of why things should remain as they were before.