Change sorting procedure to make better use of polymorphism
Most general sorting routines take a "less than or equal" comparison function as
arguement and produce a sorting function for the type being compared by it. The
stable-sorting |sort_by| however takes a map to integers or rationals as
argument instead of a comparison function; with these two polymorphically typed
forms, the many monomorphic versions of |sort_by| were removed, and had to be
because there presence makes calling the polymorphic one ambiguous. We did not
(yet) decide to make this polymorphic in the mapped-to-for-comparison type.