Change grammar to exclude parenthesised single item identifier patterns
This allowed anomalies of the kind |let ((a):b)= 3 in a+b|, where in parsing
the identifier |a| is overwritten by the |b| in the parsing process, and which
therefore emits the strange error message that the use of |a| is undefined here.

The liberty to redundantly parenthesise an identifier in a pattern was actually
used in one script, presumably by accident; the script was corrected.

Single item patterns are still allowed in tagged case discrimination clauses, as
in "case ... |no_vec:...|solution(s): ... esac", so we added a new grammatical
class |case_pattern| slightly more lenient than the now changed |closed_pattern|.

In passing a strange irregularity of the grammar was ironed out: in mentioned
case clauses, a naked |pattern| (without parentheses) was allowed before a
suffixed case tag, unless that was the first case in the list. This was changed
so that parenthesis around the introduced identifier(s) are always required.