Make the test for setting an operator symbol to a non-function great again
Once upon a time the grammar only accepted defining operator symbold globally
and only one at a time. By now operator symbols are allowed any identifier
pattern, which means they can be in many places: global and local definition
groups, and other patterns like those in case discrimination clauses.

The case where the parser signalled via an argument |overload==2| the definition
of a single operator symbol has disappeared, so the test, while present in the
code, was never activated. We replace the test by one occurring whenever an
identifier is added to a global definition group or to a local lexical |layer|.