Ensure no overload table entries ever have 0 variants, to ensure consistency
This is achieved by making the |overload_table::remove| method clean up the
table when a last overloaded instance of an identifier is removed. More than to
simplify testing for presence in the table this serves to ensure that no
behaviour can be influenced by "phantom entries" in the overload table.