Make operator casts call (second) |conform_types| after deducing the type
The context type has not been used when matching the given argument type to the
variants of the operator, so once it is found we should treat it in the same way
if an applied identifier had been looked up: conform to the context type, which
may involve (further) unification and/or specialising the context type.

We made a similar modification for |$|, the expression giving the last value
produced, which should be handled as applied identifier, but was overlooked.