Make '<' and '>' scan as themselves (character code tokens) rather than OPERATOR
The special scanning of these tokens at the initial position in a command is not
changed (they scan as FROMFILE, FORCEFROMFILE or TOFILE), nor is their possible
combination with a following '=', but in the remaining cases the conversion from
'<' or '>' to the |operator| category is now done in the parser. In passing we
did remove scanning of |<:=| as |OPERATOR_BECOMES| which semms quite useless.

This is in preparation for the introduction of user defined type constructores,
where we intend to use '<' and '>' as delimiters for their type-argument lists.