First implementation for definitions of type constructors
A simple type definition inside 'any_type S,T begin ... end' is interpreted as
the definition of a type constructor, with 'S' and 'T' as argument types.

Thus definition is rather clumsy, notably for the fact that expansion happens
during parsing actions that have limited access to information. But it was
tested to work with definition 'set_type either= (S fail|T success)' notably it
makes the injector functions 'fail' and 'success' properly polymorphic. The
injector tags cannot yet be used in a discrimination clause, as that requires a
heavy-weight grouped type definition, for which abstraction is not yet possible.

This commit introduces two major changes: (1) one can write a light-weight type
definition enclosed in an 'any_type' construct, which then becomes a type
constructor definition with the type variables introduced becoming type
parameters; (2) thus defined type constructors can be invoked following them by
a type list in pointy brackets, as in 'either<string,mat>'.

For (1) the function define_type_identifier from global.w now takes an extra
argument for the arity of type constructor, which is called from the parsing
action for the new type definition clause, and uses a new factory function
|type::constructor| to build a type with degree equal to that arity.

For (2), after having modified the lexer to pass '<' and '>' as tokens distinct
from OPERATOR, we also need to separate TYPE_ID from TYPE_OPERATOR in order to
avoid parsing conflicts (a TYPE_ID can be followed by '<' as operator, but after
TYPE_CONSTRUCTOR it is definitely a pointy left bracket). In order for the
parsing action 'expand_type_constructor' to know which type variables to protect
from substitution, the lexer now provides a typevar_level method. (This could be
avoided by not using |substitution| which continues recursively, but a simpler
one-level substitution function, which would not need protection of variables.)