Fix bug introduced in commit 24e57a71: used variable after moving from it
Two uses of |gamma| were made (type 1 or 2 Cayley transform), one conditional
the other unconditional. In order to eventually move from it unconditionally,
the statements must be ordered so that the conditional use comes, even if this
looks a bit unnatural. This was done correctly, until the statements themselves
were simplified, and in passing I permuted the statement to what looks like a
more natural ordering, but in doing so made the conditional use access a
variable after it had been (unconditionally) been mode from, causing a crash.