Added option to do bottom_layer with smaller Q
Option is used with small_test_flag:=true (default false). In this
case the parabolic Q defined by d_lambda (from
theta_stable_quasi_data) is replaced by (a version of) the one from
theta_stable_data.

This version is still broken in unequal rank, because the construction
of the small Q is not sufficiently theta_stable. I'll fix soon?