Corrected bugs and tried to address speed issues
Serious bugs were related to the fact that the barycenter of a global
face is not the barycenter of the corresponding local face. This is
addressed by an old function face_bary, but I failed to use it
several times. This led to incorrect results for SU(3,2). (Rep was
computed at its local barycenter, looked up in a supposedly complete
table, and not found. Incorrect conclusion was that the rep was not
unitary; actual conclusion was that it was not computed at its
global barycenter.)

Addressed the bug in master by computing KL_sum_at_s_to_ht twice, and
using only the second value. Don't know what the speed cost is.

I don't understand the speed issues.