Undo sloppy code rewriting in commit b6d2538ca, which broke |lookup_full_block|
In fact the functions being called like |make_relative_to| have changed their
interface since that commit, so the change here consists of making the two
branches if the main conditional in |lookup_full_block| very similar again,
apart from the |add_block| call and looking up the reduced parameter in the
resulting block. This makes it very tempting to change the code again so that
this difference is the only conditional part, as it was before mentioned commit.

An attempt to answer why, in contrast to what is done here, there is in
|Rep_table::lookup| no need to call |make_relative_to| when a fresh block is
built. It seems that with no previous block present, the block construction must
reproduce the block element for which a (partial) block is being created as-is.
In the case of a full block on the other hand, the block might contain another
element, related to it by a symmetry of the block and therefore equivalent for
looking-up purposes (but not as parameter), and this element could be returned
from the call to |block.representative|, but not match |srm|. But maybe this is
false, and then |lookup| must be modified just as |lookup_full_block| was here.