Adapt to compiler getting noisy about deprecated |unary_predicate| and friend
Apparently it is no longer necessary to be able to interrogate a predicate type
about the type(s) to which it applies, so that inheritance from either
|unary_predicate| or |binary_predicate| as a means to set |argument_type| as
embedded typedef is predicated. But our code does use this in one place, so
that we now just set such a typedef manually where it is needed.