Further progress on endoscopic lifting
Main change: given a parameter p_H for the endoscopic group H (a RealForm)
get y a KGBElt_gen for the dual innner class to H (Note: y is NOT
a KGBElt for dual(p_H).real_form, which has underlying root datum the integrality datum
so y does not define a valid RealForm. This is OK, it doesn't need to, it only needs
to define an InnerClass. This required some cleaning up of the definitions