Completely redefine |add_typedefs|; recursive definitions are no longer compared
The current code has the ambition to work for user type constructor definitions,
even though the grammar does not yet allow |any_type| constructs to contain
grouped type definitions. Since comparing recursive definitions would also have
to deal with detecting the potential of equality between applications of
recursive type constructors, which the old algorithm is not really up to, it was
decided to replace it by a simpler mechanism that abandons the ambition to
detect equivalence among recursive type definitions: recursive types, whether
directly defined or by application of a recursive type constructor, are only
equivalent if constructed using the same recursive definition and with
equivalent type arguments in the case of type constructors. For the non
recursive case the definition of type equivalence remains structural: type
definitions are expanded and equivalence is identity of the result.

Work remains to be done in global.w (|process_type_definitions|) to prepare for
grouped type definitions to be adapted to type constructor definitions.