Allow script branch everywhere
Rearranged a lot so that the script versions of branch defined in
FPP_faces_herm.at could always be used (if my_branch_flag is true,
which is the default). Output from FPP now includes a line saying
how much time was spent in built-in branch and in script branch.