Correct a mistake (not bug: no wrong answers) found by Jeff
The code uses "best_Dirac_bound" to look for a better K-type than the
LKT to get a Dirac inequality estimate. Finding these better K-types tau
is cheap, but APPLYING the corresponding Dirac inequality requires
branching to height(tau). If tau is gotten by too many steps, this
is a big height difference, and the branching is slow.

The "better taus" are stored in a hash table indexed by mu in
Dirac.at. More or less they were constrained to number of steps
equal to Dirac_best_steps. But the construction of better taus
is recursive; so if in the process I found a tau0 whose better
taus had already been computed, then I just added the list of better
ones for tau0 to my list for mu ending at tau0.

In this process, I forgot to cap the list using best_Dirac_steps.

The result was that the hash table sometimes contained a longer list
of better K-types than it was supposed to. When such a case was
encountered, the best_Dirac_bound Dirac inequality was (correctly)
applied, but doing more branching work than was intended. Therefore
the computation slowed down.

This case could only arise if best_Dirac_bound had previously been
applied to some K-type above the LKT under consideration (given by
(x,lambda)). So it never happened with a run on a single (x,lambda),
and only rarely on few (x,lambda). But Murphy's law guaranteed that
on Jeff's jobs, each seeing a few thousand different (x,lambda), there
would eventually be slowness.

Fixed, I think; I'm testing now.