Make the built-in |K_type_formula| memoize its results
The non-memoizing version is still available as |K_type_formula_raw|.

Even if a previously stored value is used, the new built-in spends some time
producing the result, since a |KTypePol| user value needs to be constructed from
the stored reference that the library function almost immediately returns (the
same is true, and always has been, for deformation formulas). Here constructing
the result involves converting integer coefficients to |Split|, and truncating
the polynomial in case a lower height is requested than the one stored.