Fix a bug in |Free_Abelian_light::const_reverse_iterator::pop|
There was a spurious decrement of a vector iterator in a case where the
recursive call of |pop| had already done a decrement (resulting in the removal
of the vector in which that decrement was done). This bug was undiscovered due
to reverse iteration being rare; it turned up in |Free_Abelian_light::rbegin|
when a call (in the context of memoizing |K_Type_formula|) was recently added.

In passing some comments were added or corrected. The actual bug fix is in
the final line of difference for this commit.