Addressed error "Real form mismatch...
...when adding two K_types." This is because atlas has two or more
different interpretations of what it means for two RealForms to
coincide: they can test as equal, but atlas still might not allow you
to add their K-types.

Tried to make all the tables in big_unitary_hash use K-types attached
to the real forms in real_form_hash, and to return answers attached to
the real forms in the questions. But this is a mess, and I don't
understand it properly.