reworked bottom layer for speed...
...bottom_layer arguments were previously done on a maximal
(theta-stable) Levi. Now (for each bottom layer K-type tau) they
are done on the smallest Q for which tau is bottom layer. This
will have no effect on places like KGB(E8_s,220459) where the minimal
theta-stable Levi available is already maximal. But on average it
ought to help a lot.

For the first 200 KGBElts in E7_s with more_shift(5), time went down
from 1700 aeconds to 1100 seconds on my laptop.