Change |type_expr| handling: avoid using |tuple| and friends without expanding
It used to be safe to use the |component_type|, |func| and |tuple| methods after
inspecting |kind()| to see that the variant can be found, possibly after
switching to the |type_expr::type_map| entry corresponding to a tabled type.
However, now that tabled types can have argument lists that need to be
substituted into that entry, we cannot safely use the result of mentioned
methods without performing that substitution manually; this more or less defeats
the original purpose of these methods to make tabled types transparent for most
type handling. The alternative now implemented is to do the expansion before
calling these methods; this sometimes involves creating a local variable to hold
the expanded value, and at other times can be done by assigning the expansion
back to the |type_expr| variable that was tabled (if we have non |const| access
to it). A non-|const| method |expand| was added to facilitate the latter.