Speeded up localFD_1Dirac2_by_zero
Did a fair amount of testing to see what seemed fastest. I don't
understand the final condition, but it isn't entirely ridiculous, and
it seems to work all right including E8. Regrettably it is still a
little SLOWER than the best times from earlier this month; but not a lot.

Also made Qs2B_flag true by default, so that a smaller parabolic is
used as base of bottom layer.