Fix an oversight that could cause a crash while in the lexical analyzer
Due to the generated parser sometimes doing a default reduction even when an
incompatible look-ahead token is already scanned, the lexer could try to push a
type variable list when in fact no nest layer to hold them is present: this
crashed the program before the inevitable following syntax error is produced.
We now test for the condition that would cause the crash, and avoid it.