Fix wrong place for the call of |alcove_center| during (twisted) deformation
In order to be effective in depth, the potential call to |alcove_center| should
happen at each level of the recursion. This used to be the case, but in a
reorganization of the deformation computation in commit 526a47cf02, the code
that was always run upon entering |Rep_table::full_deformation| found itself
outside the recursion, which now takes place for |Rep_table::deformation|. The
code in question is now moved to the start of the latter method, which should
suffice to prevent unchecked growth of denominators. #METOO the twisted case.