Correct an error in to_no_Cminus...
...versions introduced in bottom_layer.at. The error affects only the
unequal rank case, and only when small_test_flag is true. It appears
just to cause the script to stop (rather than producing bad
answers. The error is old, and I thought I had run E6_s several times
in the interim; but perhaps not.

I am slightly concerned that the new theta_stable_data_small@Param
might take a hermitian p to a non-hermitian pL. I don't know that that
happens, and if it does it ought to be harmless... code will just be a
bit less efficient than intended.