Add general |type::unify| method, preparing for using |type| in |convert_expr|
When we will replace the |type_expr| argument of |convert_expr| by a |type|, to
allow for the expected type to be given by any polymorphic type (meaning that
some instance of it should be found) rather than just by an expression with
undetermined parts as it is currently, we shall need to replace calls of
|specialise| by calls of |unify|. In order to make this transition less
painfull, we prepare the |unify| method to be able to handle undetermined parts
as if they were freah polymorphic variables. Maybe this is not necessary because
|type::wrap| already makes this change manifest, but we can always simplify
later if this turns out to be the case.

In passing we also reorganised |type_assignment| by moving the functions
|substitute| to become a method, and adding a method |unify|. The later might
replace the |can_unify| function (or maybe it too can be moved to become a
method of |type_assignment|, if having |const| arguments remains important).