no substantial changes
The script edges_F4_E7.at contains precomputed data about edges in F4,
E6, and E7. The modification in this commit loads quietly, but is
otherwise identical.

Added a function kill_for_big_uTime@void in FPP_global_Dirac.at, which
writes a suicide note if

a) kill_for_big_uTime_flag is true (default false), and
b) the time so far spent in is_unitary exceeds a positive bound
   kill_for_big_uTime_bound (default -1)

A function activate_kill_for_big_uTime@int makes the flag true and
sets the bound to the int.

If the defaults are unchanged, then there is no change at all. If the
kill is activated, the suicide note may be printed, but there is no
other effect; the idea is that a supervising program can read the note
and kill the atlas process.