Cleaned up is_unitary_test_big_SIMPLE_interrupt
This is almost entirely speed modifications. Still somewhat slower
than not using it in in almost all rank up to 6 cases, but it almost
always reduces the time in is_unitary. For that reason I hope it will
be useful in E7 and E8.

Scripts now print time spent in up_mu at end of a job (after the
reports on various kinds of is_unitary).

The new function is still invoked by up_mu_flag :=true (default false).