used short_shifts also in DiracBestTest
short_shifts makes it easy to increase Dirac_best_steps, and so to get
better bounds on nu. For now only the SMALLEST bound is used, and the
K-multiplicity requirement may fail often for that; so it's possible
to find FEWER Dirac_best_elims. Should be easy to address that.