Cleaned up is_unitary_test_big_SIMPLE_interrupt
The function up_mu_hash.up_heights did not behave as expected. One of
the many little issues was that the built-in Cayley(nci root, p)
did not behave as I expected; we'll need to discuss whether there is
an issue. (I expected p.x to become Cayley(nci root, p.x). But this
does not happen for some standard but not final Param's.)